+++ packages/core/src/components/CssBaseline/CssBaseline.styled.tsx
@@ -18,6 +18,7 @@ export const CssBaseline = createGlobalStyle`
         height: 100%;
         width: 100%;
         font-size: 62.5%;
+        scroll-behavior: smooth;
     }
     body {
         margin: 0;