+++ packages/forms/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.30.5](https://github.com/medly/medly-components/compare/@medly-components/forms@1.30.4...@medly-components/forms@1.30.5) (2021-09-20)
+
+**Note:** Version bump only for package @medly-components/forms
+
+
+
+
+
+## [1.30.4](https://github.com/medly/medly-components/compare/@medly-components/forms@1.30.3...@medly-components/forms@1.30.4) (2021-09-20)
+
+**Note:** Version bump only for package @medly-components/forms
+
+
+
+
+
+## [1.30.3](https://github.com/medly/medly-components/compare/@medly-components/forms@1.30.2...@medly-components/forms@1.30.3) (2021-09-19)
+
+**Note:** Version bump only for package @medly-components/forms
+
+
+
+
+
+## [1.30.2](https://github.com/medly/medly-components/compare/@medly-components/forms@1.30.1...@medly-components/forms@1.30.2) (2021-09-18)
+
+**Note:** Version bump only for package @medly-components/forms
+
+
+
+
+
+## [1.30.1](https://github.com/medly/medly-components/compare/@medly-components/forms@1.30.0...@medly-components/forms@1.30.1) (2021-09-17)
+
+**Note:** Version bump only for package @medly-components/forms
+
+
+
+
+
+# [1.30.0](https://github.com/medly/medly-components/compare/@medly-components/forms@1.29.15...@medly-components/forms@1.30.0) (2021-09-16)
+
+
+### Features
+
+* **core:** textfield custom active border styles ([#424](https://github.com/medly/medly-components/issues/424)) ([2a22fc7](https://github.com/medly/medly-components/commit/2a22fc7bf85a8f4446ad0d837403044f6b194693))
+
+
+
+
+
 ## [1.29.15](https://github.com/medly/medly-components/compare/@medly-components/forms@1.29.14...@medly-components/forms@1.29.15) (2021-09-15)
 
 **Note:** Version bump only for package @medly-components/forms