+++ packages/layout/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.29.24](https://github.com/medly/medly-components/compare/@medly-components/layout@1.29.23...@medly-components/layout@1.29.24) (2021-09-20)
+
+**Note:** Version bump only for package @medly-components/layout
+
+
+
+
+
+## [1.29.23](https://github.com/medly/medly-components/compare/@medly-components/layout@1.29.22...@medly-components/layout@1.29.23) (2021-09-20)
+
+**Note:** Version bump only for package @medly-components/layout
+
+
+
+
+
+## [1.29.22](https://github.com/medly/medly-components/compare/@medly-components/layout@1.29.21...@medly-components/layout@1.29.22) (2021-09-19)
+
+**Note:** Version bump only for package @medly-components/layout
+
+
+
+
+
+## [1.29.21](https://github.com/medly/medly-components/compare/@medly-components/layout@1.29.20...@medly-components/layout@1.29.21) (2021-09-18)
+
+**Note:** Version bump only for package @medly-components/layout
+
+
+
+
+
+## [1.29.20](https://github.com/medly/medly-components/compare/@medly-components/layout@1.29.19...@medly-components/layout@1.29.20) (2021-09-17)
+
+**Note:** Version bump only for package @medly-components/layout
+
+
+
+
+
+## [1.29.19](https://github.com/medly/medly-components/compare/@medly-components/layout@1.29.18...@medly-components/layout@1.29.19) (2021-09-16)
+
+**Note:** Version bump only for package @medly-components/layout
+
+
+
+
+
 ## [1.29.18](https://github.com/medly/medly-components/compare/@medly-components/layout@1.29.17...@medly-components/layout@1.29.18) (2021-09-15)